packages/media/pyproject.toml
[project]
name = "forensic-media"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "imagehash>=4.3.2",
    "pillow>=12.1.1",
    "psycopg2-binary>=2.9.11",
    "sqlalchemy>=2.0.48",
]
